FIR/UAST: commonize lambda parameter test
