+++ plugins/kotlin/uast/uast-kotlin/test/org/jetbrains/uast/test/comparasion/FE1UastApiTest.kt
@@ -165,6 +165,11 @@ class FE1UastApiTest : AbstractFE1UastTest() {
             doTest("ReifiedParameters", ::checkCallbackForReifiedParameters)
         }
 
+        @TestMetadata("LambdaParameters.kt")
+        fun testLambdaParameters() {
+            doTest("LambdaParameters", ::checkCallbackForLambdaParameters)
+        }
+
         @TestMetadata("SAM.kt")
         fun testSAM() {
             doTest("SAM", ::checkCallbackForSAM)
